agregando tablas y servicios

+++ src/main/java/Clases/Etiqueta.java
@@ -16,6 +16,6 @@ public class Etiqueta {
     public long getId() { return id; }
     public String getEtiqueta() { return etiqueta; }
 
-    public void setId(long id) { this.id = id; }
+    public void setId(int id) { this.id = id; }
     public void setEtiqueta(String etiqueta) { this.etiqueta = etiqueta; }
 }
